requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
qrcode[pil]==7.4.2
python-dotenv==1.0.1
Pillow>=11.0.0
reportlab>=4.0.0
